# LibUnicode: Perform code point case conversion lookups in constant time