First commit is a compile-time performance improvement to cut the runtime of GenerateUnicodeData from
3.4 seconds to 1.2 seconds.

Next is replacing the lookup tables for case information (which we currently binary-search at runtime) with 2-stage, constant-time lookup tables. Not *as* much of a win as code point property tables, but still a measurable improvement, and it's nice to have these tables all behave the same:

In total, this change:

* Does not change the size of libunicode.so (which is nice because,
  generally, the 2-stage lookup tables are expected to trade a bit
  of size for performance).

* Reduces the runtime of the new benchmark test case added here from
  1.383s to 1.127s (about an 18.5% improvement).